Tahiti 2014.

Dataset includes weekly laboratory confirmed Zika cases from Tahiti in French Polynesia, during an outbreak in 2013 and 2014 as well as monthly average temperature data and rainfall for Tahiti during this period

Usage

tahiti_2013

Format

An object of class tbl_df (inherits from tbl, data.frame) with 25 rows and 6 columns.

Details

#' A data frame with 25 rows and 6 columns:

date

Date of epidemiological week

cases

Number of laboratory confirmed cases in Tahiti

reporting_sites

Number of sentinel sites in Tahiti reporting that month

total_sites

Total sentinel sites Tahiti

av_temp

Average monthly temperature for Tahiti, in °C

av_temp

Average monthly rainfall for Tahiti, in mm

...

#' @source https://journals.plos.org/plosntds/article?id=10.1371/journal.pntd.0004726